requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
email-validator==2.1.0
imaplib2==3.6
# smtplib is part of Python standard library
openai==1.3.7
nltk==3.8.1
textblob==0.17.1
schedule==1.2.0
requests==2.31.0
python-telegram-bot==20.7
slack-sdk==3.26.1
gTTS==2.4.0
pydub==0.25.1
redis==5.0.1
celery==5.3.4
jinja2==3.1.2
aiofiles==23.2.1 